+++ matrix-sdk-android/src/main/java/org/matrix/android/sdk/internal/auth/data/WhoAmIResponse.kt
@@ -1,0 +1,17 @@
+/*
+ * Copyright 2026 Voyage Client
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only
+ * Please see LICENSE files in the repository root for full details.
+ */
+
+package org.matrix.android.sdk.internal.auth.data
+
+import com.squareup.moshi.Json
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+internal data class WhoAmIResponse(
+        @Json(name = "user_id") val userId: String,
+        @Json(name = "device_id") val deviceId: String? = null
+)